frontend: move QueryURL from _listing to _base
