+++ main/local_modules/frontend_base/controllers/base.py
@@ -7,9 +7,31 @@ from openerp.addons.web import http
 from openerp.addons.website.controllers.main import Website
 from openerp.addons.website.models.website import hashlib
 
-from openerp.http import request
+from openerp.http import request, werkzeug
 
 _logger = logging.getLogger(__name__)
+
+
+class QueryURL(object):
+    def __init__(self, path='', **args):
+        self.path = path
+        self.args = args
+
+    def __call__(self, path=None, **kw):
+        if not path:
+            path = self.path
+        for k, v in self.args.items():
+            kw.setdefault(k, v)
+        l = []
+        for k, v in kw.items():
+            if v:
+                if isinstance(v, list) or isinstance(v, set):
+                    l.append(werkzeug.url_encode([(k, i) for i in v]))
+                else:
+                    l.append(werkzeug.url_encode([(k, v)]))
+        if l:
+            path += '?' + '&'.join(l)
+        return path
 
 
 def small_image_url(record, field):
